Refactor modules and add configurations
API module was removed and POM file is modified
application.properties are modified for DB configurations
