--- api/src/main/java/org/sefglobal/scholarx/ScholarxApplication.java
@@ -1,13 +1,0 @@
-package org.sefglobal.scholarx;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ScholarxApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ScholarxApplication.class, args);
-	}
-
-}
