--- api/src/main/java/org/sefglobal/scholarx/ServletInitializer.java
@@ -1,13 +1,0 @@
-package org.sefglobal.scholarx;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(ScholarxApplication.class);
-	}
-
-}
