+++ src/main/java/org/sefglobal/scholarx/ScholarxApplication.java
@@ -1,0 +1,21 @@
+package org.sefglobal.scholarx;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class ScholarxApplication extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(ScholarxApplication.class);
+	}
+
+	public static void main(String[] args) {
+		SpringApplication.run(ScholarxApplication.class, args);
+	}
+}
