+++ src/main/resources/application.properties
@@ -1,0 +1,9 @@
+## Database Properties
+spring.datasource.url = jdbc:mysql://127.0.0.1:3306/${DB_NAME}?useSSL=false&useUnicode=true&characterEncoding=UTF-8
+spring.datasource.username = ${DB_USERNAME}
+spring.datasource.password = ${DB_PASSWORD}
+## Hibernate Properties
+# The SQL dialect makes Hibernate generate better SQL for the chosen database
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
+endpoints.jmx.unique-names = true
+spring.jpa.hibernate.ddl-auto = update
